Fix misplaced ...ListFragment spread.
https://github.com/apollographql/apollo-client/pull/8782#discussion_r797984651

Previously, the ListFragment was effectively ignored because its type
condition (on Query) did not match the "Item" __typename.

Although the code was definitely wrong before, this change ends up not
making a visible difference because the Item.id field provided by
ListFragment is the default field used for normalization, so it gets
picked up and used regardless of whether it's mentioned in the query
(even when ListFragment is completely ignored, as it was before).

To make this change matter in a visible way, I added an extra root query
field to ListFragment to strengthen the tests.